Refactor course, department, and professor models by removing redundant classes and enhancing structure
